chore: move addons configuration to custom module
